+++ api/src/main/java/org/grnet/cat/api/endpoints/AssessmentsV2Endpoint.java
@@ -25,6 +25,7 @@ import org.grnet.cat.api.utils.CatServiceUriInfo;
 import org.grnet.cat.constraints.NotFoundEntity;
 import org.grnet.cat.dtos.InformativeResponse;
 import org.grnet.cat.dtos.assessment.ZenodoAssessmentInfoResponse;
+import org.grnet.cat.dtos.assessment.zenodo.ZenodoDepositResponse;
 import org.grnet.cat.dtos.assessment.registry.AdminJsonRegistryAssessmentResponse;
 import org.grnet.cat.dtos.assessment.registry.JsonRegistryAssessmentRequest;
 import org.grnet.cat.dtos.assessment.registry.UserJsonRegistryAssessmentResponse;
@@ -36,9 +37,7 @@ import org.grnet.cat.services.zenodo.ZenodoService;
 import org.grnet.cat.utils.Utility;
 
 import java.io.IOException;
-import java.util.Map;
 import java.util.concurrent.ExecutionException;
-import java.util.concurrent.CompletableFuture;
 
 import static org.eclipse.microprofile.openapi.annotations.enums.ParameterIn.QUERY;
 
@@ -638,17 +637,16 @@ public class AssessmentsV2Endpoint {
         return Response.ok().entity(response).build();
     }
 
-
-    @Tag(name = "Assessment")
+@Tag(name = "Zenodo")
     @Operation(
             summary = "Zenodo publish an assessment.",
             description = "Allows a user to publish an assessment to zenodo.")
     @APIResponse(
             responseCode = "200",
-            description = "Assessment published successfully.",
+            description = "Request to publish assessment to zenodo submitted  successfully.",
             content = @Content(schema = @Schema(
                     type = SchemaType.OBJECT,
-                    implementation = ZenodoAssessmentInfoResponse.class)))
+                    implementation = InformativeResponse.class)))
     @APIResponse(
             responseCode = "400",
             description = "Invalid request payload.",
@@ -681,37 +679,38 @@ public class AssessmentsV2Endpoint {
                     implementation = InformativeResponse.class)))
     @SecurityRequirement(name = "Authentication")
     @POST
-    @Path("/zenodo/publish/{id}")
+    @Path("/zenodo/publish/{assessment-id}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_OCTET_STREAM) // This consumes binary data
     @Registration
-    public Response createZenodoDeposit(@Parameter(
-                                                description = "The ID of the assessment to publish to zenodo.",
-                                                required = true,
-                                                example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
-                                                schema = @Schema(type = SchemaType.STRING)) @PathParam("id")
-                                        @Valid @NotFoundEntity(repository = MotivationAssessmentRepository.class, message = "There is no Assessment with the following id:") String id,
+    public Response publishZenodoAssessment(@Parameter(
+                                                    description = "The ID of the assessment to publish to zenodo.",
+                                                    required = true,
+                                                    example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
+                                                    schema = @Schema(type = SchemaType.STRING)) @PathParam("assessment-id")
+                                            @Valid @NotFoundEntity(repository = MotivationAssessmentRepository.class, message = "There is no Assessment with the following id:") String assessmentId,
 
-                                        byte[] pdfFile) throws IOException, InterruptedException, ExecutionException {
+                                            byte[] pdfFile) {
 
         if (!isValidPdf(pdfFile)) {
             return Response.status(Response.Status.BAD_REQUEST)
                     .entity("Invalid PDF file format.")
                     .build();
         }
-       var message= zenodoService.publishAssessment(id, pdfFile,utility.getUserUniqueIdentifier());
-        var response=new InformativeResponse();
-        response.code=202;
-        response.message=  message;
+        var message = zenodoService.publishAssessment(assessmentId, pdfFile, utility.getUserUniqueIdentifier());
+        var response = new InformativeResponse();
+        response.code = 202;
+        response.message = message;
         return Response.ok(response).build();
     }
-    @Tag(name = "Assessment")
+
+    @Tag(name = "Zenodo")
     @Operation(
             summary = "Get a Zenodo  assessment.",
-            description = "Allows a user to view an assessment in zenodo.")
+            description = "Allows a user to view information of an assessment in zenodo, as it is stored to CAT service.")
     @APIResponse(
             responseCode = "200",
-            description = "Assessment published successfully.",
+            description = "Information of zenodo assessment, in CAT service, displayed successfully.",
             content = @Content(schema = @Schema(
                     type = SchemaType.OBJECT,
                     implementation = InformativeResponse.class)))
@@ -747,31 +746,30 @@ public class AssessmentsV2Endpoint {
                     implementation = InformativeResponse.class)))
     @SecurityRequirement(name = "Authentication")
     @GET
-    @Path("/zenodo/assessment/{id}")
+    @Path("/zenodo/{assessment-id}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON) // This consumes binary data
     @Registration
     public Response getZenodoAssessment(@Parameter(
-                                                description = "The ID of the assessment uploaded to zenodo.",
-                                                required = true,
-                                                example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
-                                                schema = @Schema(type = SchemaType.STRING)) @PathParam("id")
-                                        @Valid @NotFoundEntity(repository = MotivationAssessmentRepository.class, message = "There is no Assessment with the following id:") String id) throws IOException, InterruptedException, ExecutionException {
-     var response=   zenodoService.getAssessment(id);
+            description = "The ID of the assessment uploaded to zenodo.",
+            required = true,
+            example = "c242e43f-9869-4fb0-b881-631bc5746ec0",
+            schema = @Schema(type = SchemaType.STRING)) @PathParam("assessment-id")
+                                        @Valid @NotFoundEntity(repository = MotivationAssessmentRepository.class, message = "There is no Assessment with the following id:") String assessmentId) throws IOException, InterruptedException {
+        var response = zenodoService.getAssessment(assessmentId);
         return Response.ok().entity(response).build();
     }
 
-
-    @Tag(name = "Assessment")
+    @Tag(name = "Zenodo")
     @Operation(
-            summary = "Publish a Zenodo  deposit.",
-            description = "Allows a user to publish a specific deposit.")
+            summary = "Get Info about a depoisit in Zenodo.",
+            description = "Allows a user to view deposit info  existing in zenodo, as the info exist in the CAT service.")
     @APIResponse(
             responseCode = "200",
-            description = "Assessment published successfully.",
+            description = "Zenodo deposit information displayed successfully.",
             content = @Content(schema = @Schema(
                     type = SchemaType.OBJECT,
-                    implementation = InformativeResponse.class)))
+                    implementation = ZenodoDepositResponse.class)))
     @APIResponse(
             responseCode = "400",
             description = "Invalid request payload.",
@@ -803,24 +801,82 @@ public class AssessmentsV2Endpoint {
                     type = SchemaType.OBJECT,
                     implementation = InformativeResponse.class)))
     @SecurityRequirement(name = "Authentication")
-    @POST
-    @Path("/zenodo/deposit/{id}")
+    @GET
+    @Path("/zenodo/deposit/{deposit-id}")
     @Produces(MediaType.APPLICATION_JSON)
-    @Consumes(MediaType.APPLICATION_JSON) // This consumes binary data
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Registration
+    public Response getZenodoDeposit(@Parameter(
+            description = "The ID of the deposit existing in zenodo.",
+            required = true,
+            example = "13467",
+            schema = @Schema(type = SchemaType.STRING)) @PathParam("deposit-id")
+                                     String depositId) {
+        var response = zenodoService.getDeposit(depositId, utility.getUserUniqueIdentifier());
+        return Response.ok().entity(response).build();
+    }
+
+    @Tag(name = "Zenodo")
+    @Operation(
+            summary = "Publish a Zenodo  deposit.",
+            description = "Allows a user to publish a specific deposit to Zenodo.")
+    @APIResponse(
+            responseCode = "200",
+            description = "Assessment published successfully.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "400",
+            description = "Invalid request payload.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "401",
+            description = "User has not been authenticated.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "403",
+            description = "Not permitted.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "404",
+            description = "Deposit not found.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @APIResponse(
+            responseCode = "500",
+            description = "Internal Server Error.",
+            content = @Content(schema = @Schema(
+                    type = SchemaType.OBJECT,
+                    implementation = InformativeResponse.class)))
+    @SecurityRequirement(name = "Authentication")
+    @POST
+    @Path("/zenodo/publish/deposit/{deposit-id}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Consumes(MediaType.APPLICATION_JSON)
     @Registration
     public Response publishZenodoDeposit(@Parameter(
             description = "The ID of the deposit to publish to Zenodo",
             required = true,
             example = "183367",
-            schema = @Schema(type = SchemaType.STRING)) @PathParam("id")
-                                         String id) throws IOException, InterruptedException, ExecutionException {
-          zenodoService.publishDepositToZenodo(id,utility.getUserUniqueIdentifier());
-        var response=new InformativeResponse();
-        response.code=202;
-        response.message="Process to publish Deposit with ID: "+id+" is in progress... You will be informed via email when completed.";
+            schema = @Schema(type = SchemaType.STRING)) @PathParam("deposit-id") String depositId) {
+
+        zenodoService.publishDepositToZenodo(depositId, utility.getUserUniqueIdentifier());
+
+        var response = new InformativeResponse();
+        response.code = 202;
+        response.message = "Process to publish Deposit with ID: " + depositId + " is in progress... You will be informed via email when completed.";
 
         return Response.ok().entity(response).build();
     }
+
     private boolean isValidPdf(byte[] fileContent) {
         // Check the magic number for PDF files: %PDF- (hex: 0x25 0x50 0x44 0x46)
         if (fileContent.length < 4) {